fix(projects): address code review findings from PR #2057
Prevent shell injection by replacing exec with execFile for git
operations, make WorkspaceProjectAnalysis.description optional to match
WorkspaceProjectInfo, and extract duplicated git URL helpers into a
shared git-url-utils module.
